requirements.txt
asgiref==3.4.1
attrs==24.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.28.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.0.1
oauthlib==3.2.2
packaging==21.0
Pillow==8.3.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==1.0.1
pytz==2021.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.26.0
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.7.2
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.12.2
uritemplate==3.0.1
urllib3==1.26.6
views==0.3
